basic_environment/requirements.txt
stable-baselines3==1.1.0
gym==0.17.3
torch==1.8.1
tensorflow==2.12.0
google-cloud-storage
tensorboard
opencv-python
numpy
